# Resolver perf: Optimise package exports `flattenLegacySubpathValues` (1/n)